cleanup: delete LocksRPCService — lock RPC now via kernel syscalls
Lock operations (sys_lock, sys_unlock, lock_info, lock_list,
lock_extend, lock_force_release, lock_release) are @rpc_expose
on NexusFS — auto-discovered by NexusVFSService Call() RPC.

- Delete src/nexus/server/rpc/services/locks_rpc.py
- Remove LocksRPCService registration from fastapi_server.py
- Add lock_release() on NexusFS (dispatches to sys_unlock or
  lock_force_release based on --force flag, CLI compat)

EventsService retained for watch (wait_for_changes) + locked()
context manager. Full EventsService deletion is follow-up scope.
